Fix | Pagination

+++ Controllers/BookController.cs
@@ -34,7 +34,7 @@ namespace BookApi.Controllers
 
                 List<BookRepository> booksRepo = this.bookApplication.GetList(query.Search, query.Page, query.PerPage);
                 int count = this.bookApplication.Count(query.Search);
-                decimal pageInCount = count / query.PerPage;
+                decimal pageInCount = ((decimal)count) / query.PerPage;
                 PaginationModel paginate = new PaginationModel()
                 {
                     TotalPage = (int)Math.Ceiling(pageInCount),
